Initial implementation: vendor extensions -> google.protobuf.Any
NOTE: Code is still not cleaned up. This is mainly to demonstrate the idea.

Individual handlers for vendor extensions can be created using openapivendorextc, passing it the json schema and few more option.
The generted handler is in installed and is invoked by the openapic whenever it encounters a vendor extension that matches the extension name that any custom handler has registered for.

Code to look for:
How to install a sample vendor extension to proto converter
- Look into openapivendorextc/sample/INSTALL_SAMPLE.sh to see how to generate a vendor extension handler.
- Look into openapivendorextc/main.go to see how individual handlers are generated for given json schema.
How openapic invoked these handlers
- Look into changes in the OpenAPIv2/OpenAPIv2.go to see how the compiler invokes the these handlers to vendor extensions.
How to configure openapic to know about these handlers
- Look into openapic/compiler.go
What is the contract proto between the openapic and the vendorextension handlers.
- Look into vendorextension.vendorextension.proto